+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -79,7 +79,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-01-13T10:17:26.970Z")
+    created: Optional[datetime] = parse_datetime("2026-01-13T10:29:23.759Z")
     r"""The date and time the resource was created"""