+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KEVDRRW5GAVX1F7P1WDMRWNP"
+        "01KEVEEQTE2MSA83D8K560QDY1"
     )
     r"""The unique identifier of the data source"""