+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -225,7 +225,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kevdrret43q1bxjc2q3g4hpn"
+        "conv_01keveeqcp1d8rcak65decbkkb"
     )
     r"""Unique conversation identifier with `conv_` prefix."""